+++ src/commands/compile.ts
@@ -2,7 +2,6 @@ import vscode = require("vscode");
 import { isText } from "istextorbinary";
 import { AtelierAPI } from "../api";
 import {
-  config,
   documentContentProvider,
   FILESYSTEM_SCHEMA,
   OBJECTSCRIPT_FILE_SCHEMA,
@@ -14,7 +13,6 @@ import {
 import { DocumentContentProvider } from "../providers/DocumentContentProvider";
 import {
   base64EncodeContent,
-  classNameRegex,
   compileErrorMsg,
   cspAppsForUri,
   CurrentBinaryFile,
@@ -22,8 +20,9 @@ import {
   currentFileFromContent,
   CurrentTextFile,
   exportedUris,
-  getWsFolder,
+  getWsServerConnection,
   handleError,
+  isClass,
   isClassDeployed,
   isClassOrRtn,
   isCompilable,
@@ -32,11 +31,11 @@ import {
   notNull,
   outputChannel,
   RateLimiter,
-  routineNameTypeRegex,
 } from "../utils";
 import { StudioActions } from "./studio";
 import { NodeBase, PackageNode, RootNode } from "../explorer/nodes";
 import { getUrisForDocument, updateIndex } from "../utils/documentIndex";
+import { Document } from "../api/atelier";
 
 const compileWaiters = new Map<string, Set<() => void>>();
 
@@ -76,13 +75,6 @@ export async function waitForCompileToFinish(document: vscode.TextDocument, time
   });
 }
 
-async function compileFlags(): Promise<string> {
-  const defaultFlags = config().compileFlags;
-  return vscode.window.showInputBox({
-    prompt: "Compilation flags",
-    value: defaultFlags,
-  });
-}
 
 /**
  * For files being locally edited, get and return its mtime timestamp from workspace-state cache if present there,
@@ -126,8 +118,10 @@ export async function checkChangedOnServer(file: CurrentTextFile | CurrentBinary
   return mtime;
 }
 
+// Synchronize the client version and the server version of the same file
 export async function importFile(
   file: CurrentTextFile | CurrentBinaryFile,
+  willCompile: boolean,
   ignoreConflict?: boolean,
   skipDeplCheck = false
 ): Promise<any> {
@@ -160,8 +154,8 @@ export async function importFile(
     mtime < 0 ||
     (notIsfs(file.uri) &&
       vscode.workspace.getConfiguration("objectscript", file.uri).get<boolean>("overwriteServerChanges"));
-  return api
-    .putDoc(
+  try {
+    const data = await api.putDoc(
       file.name,
       {
         content,
@@ -169,69 +163,79 @@ export async function importFile(
         mtime,
       },
       ignoreConflict
-    )
-    .then((data) => {
-      // Update cache entry
-      workspaceState.update(`${file.uniqueId}:mtime`, Number(new Date(data.result.ts + "Z")));
+    );
+    // Update cache entry
+    workspaceState.update(`${file.uniqueId}:mtime`, Number(new Date(data.result.ts + "Z")));
 
-      // In case another extension has used an 'objectscript://' uri to load a document read-only from the server,
-      // make it reload with what we just imported to the server.
-      const serverUri = DocumentContentProvider.getUri(
-        file.name,
-        file.workspaceFolder,
-        undefined,
-        false,
-        undefined,
-        true
-      );
-      documentContentProvider.update(serverUri.with({ scheme: OBJECTSCRIPT_FILE_SCHEMA }));
-    })
-    .catch((error) => {
-      if (error?.statusCode == 409) {
-        const choices: string[] = [];
-        if (!enc) {
-          choices.push("Compare");
-        }
-        choices.push("Overwrite on Server", "Pull Server Changes", "Cancel");
-        return vscode.window
-          .showErrorMessage(
-            `Failed to import '${file.name}': The version of the file on the server is newer.
-What do you want to do?`,
-            ...choices
-          )
-          .then((action) => {
-            switch (action) {
-              case "Compare":
-                return vscode.commands
-                  .executeCommand(
-                    "vscode.diff",
-                    vscode.Uri.file(file.name).with({
-                      scheme: OBJECTSCRIPT_FILE_SCHEMA,
-                      authority: file.workspaceFolder,
-                      query: file.name.includes("/") ? "csp" : "",
-                    }),
-                    file.uri,
-                    `Server • ${file.name} ↔ Local • ${file.fileName}`
-                  )
-                  .then(() => Promise.reject());
-              case "Overwrite on Server":
-                // Clear cache entry
-                workspaceState.update(`${file.uniqueId}:mtime`, undefined);
-                // Overwrite
-                return importFile(file, true, true);
-              case "Pull Server Changes":
-                loadChanges([file]);
-                return Promise.reject();
-              case "Cancel":
-                return Promise.reject();
-            }
-            return Promise.reject();
-          });
-      } else {
-        handleError(error, `Failed to save file '${file.name}' on the server.`);
-        return Promise.reject();
+    if (!willCompile && isClass(file.name) && data.result.content.length) {
+      // In this case, the file must be a CLS and data.result.content must be the new Storage definitions
+      // (with the rest of the class if flags === 0)
+      const oldContent = new TextDecoder().decode(await vscode.workspace.fs.readFile(file.uri));
+      const oldContentArray = oldContent.split(/\r?\n/);
+      const storage = Buffer.isBuffer(data.result.content)
+        ? new TextDecoder().decode(data.result.content).split(/\r?\n/)
+        : data.result.content;
+      const newContentArray = updateStorage(oldContentArray, storage);
+      if (oldContentArray.some((oldLine, index) => oldLine !== newContentArray[index])) {
+        const EOL = ((<CurrentTextFile>file)?.eol ?? vscode.EndOfLine.LF) == vscode.EndOfLine.CRLF ? "\r\n" : "\n";
+        const newContent = newContentArray.join(EOL);
+        await vscode.workspace.fs.writeFile(file.uri, new TextEncoder().encode(newContent));
       }
-    });
+    }
+    // In case another extension has used an 'objectscript://' uri to load a document read-only from the server,
+    // make it reload with what we just imported to the server.
+    const serverUri = DocumentContentProvider.getUri(
+      file.name,
+      file.workspaceFolder,
+      undefined,
+      false,
+      undefined,
+      true
+    );
+    if (serverUri) documentContentProvider.update(serverUri);
+  } catch (error) {
+    if (error?.statusCode == 409) {
+      const choices: string[] = [];
+      if (!enc) {
+        choices.push("Compare");
+      }
+      choices.push("Overwrite on Server", "Pull Server Changes", "Cancel");
+      const action = await vscode.window.showErrorMessage(
+        `Failed to import '${file.name}': The version of the file on the server is newer.
+What do you want to do?`,
+        ...choices
+      );
+      switch (action) {
+        case "Compare":
+          return vscode.commands
+            .executeCommand(
+              "vscode.diff",
+              vscode.Uri.file(file.name).with({
+                scheme: OBJECTSCRIPT_FILE_SCHEMA,
+                authority: file.workspaceFolder,
+                query: file.name.includes("/") ? "csp" : "",
+              }),
+              file.uri,
+              `Server • ${file.name} ↔ Local • ${file.fileName}`
+            )
+            .then(() => Promise.reject());
+        case "Overwrite on Server":
+          // Clear cache entry
+          workspaceState.update(`${file.uniqueId}:mtime`, undefined);
+          // Overwrite
+          return importFile(file, willCompile, true, true);
+        case "Pull Server Changes":
+          loadChanges([file]);
+          return Promise.reject();
+        case "Cancel":
+          return Promise.reject();
+      }
+      return Promise.reject();
+    } else {
+      handleError(error, `Failed to save file '${file.name}' on the server.`);
+      return Promise.reject();
+    }
+  }
 }
 
 function updateOthers(others: string[], baseUri: vscode.Uri) {
@@ -250,7 +254,16 @@ function updateOthers(others: string[], baseUri: vscode.Uri) {
   });
 }
 
-export async function loadChanges(files: (CurrentTextFile | CurrentBinaryFile)[]): Promise<any> {
+/**
+ * Pull changes due to save/compile from server and write them into the local file.
+ * Also used to refresh the local copy from the server outside of the save/compile workflow.
+ * Pass `forceRefreshClasses` to replace the entire text of classes regardless of the
+ * value of the `objectscript.refreshClassesOnSync` setting.
+ */
+export async function loadChanges(
+  files: (CurrentTextFile | CurrentBinaryFile)[],
+  forceRefreshClasses = false
+): Promise<any> {
   if (!files.length) {
     return;
   }
@@ -264,7 +277,22 @@ export async function loadChanges(files: (CurrentTextFile | CurrentBinaryFile)[]
         const mtime = Number(new Date(doc.ts + "Z"));
         workspaceState.update(`${file.uniqueId}:mtime`, mtime > 0 ? mtime : undefined);
         if (notIsfs(file.uri)) {
-          const content = await api.getDoc(file.name, file.uri).then((data) => data.result.content);
+          let content: Document["content"];
+          if (
+            !(
+              isClass(file.uri.path) &&
+              !vscode.workspace.getConfiguration("objectscript", file.uri).get("refreshClassesOnSync") &&
+              !forceRefreshClasses
+            )
+          ) {
+            content = (await api.getDoc(file.name, file.uri)).result.content;
+          } else {
+            // Insert/update the storage part of class definition.
+            content = new TextDecoder().decode(await vscode.workspace.fs.readFile(file.uri)).split(/\r?\n/);
+            let storage = (await api.getDoc(file.name, file.uri, undefined, true)).result.content;
+            storage = Buffer.isBuffer(storage) ? new TextDecoder().decode(storage).split(/\r?\n/) : storage;
+            content = updateStorage(content, storage);
+          }
           exportedUris.add(file.uri.toString()); // Set optimistically
           await vscode.workspace.fs
             .writeFile(
@@ -296,13 +324,64 @@ export async function loadChanges(files: (CurrentTextFile | CurrentBinaryFile)[]
   );
 }
 
-export async function compile(docs: (CurrentTextFile | CurrentBinaryFile)[], flagsOverride?: string): Promise<any> {
+function updateStorage(content: string[], storage: string[]): string[] {
+  const storageMap = storageToMap(storage);
+  let contentString = content.join("\n");
+  contentString = contentString
+    // update existing Storages
+    .replaceAll(/\n(\s*storage\s+(\w+)\s*{\s*)(.*?)(>\s*})/gis, (_match, beforeXML, name, _oldXML, afterXML) => {
+      let newXML = storageMap.get(name);
+      if (newXML === undefined) {
+        return "";
+      }
+      newXML = newXML.slice(0, newXML.length - 1);
+      storageMap.delete(name);
+      return "\n" + beforeXML + newXML + afterXML;
+    });
+  contentString = contentString
+    // insert remaining Storages
+    .replace(/}\s*$/, (m) => {
+      for (const [name, content] of storageMap.entries()) {
+        m = `Storage ${name}\n{\n${content}\n}\n\n${m}`;
+      }
+      return m;
+    });
+  return contentString.split("\n");
+}
+
+function storageToMap(storage: string[]): Map<string, string> {
+  const map: Map<string, string> = new Map();
+  let k: string;
+  let v = [];
+  for (const line of storage) {
+    if (line.startsWith("Storage ")) {
+      k = line.slice("Storage ".length, line.length);
+      v = [];
+    } else if (k !== undefined) {
+      if (line === "{") {
+        continue;
+      } else if (line === "}") {
+        map.set(k, v.join("\n"));
+        k = undefined;
+      } else {
+        v.push(line);
+      }
+    }
+  }
+  return map;
+}
+
+export async function compile(docs: (CurrentTextFile | CurrentBinaryFile)[], askFlags = false): Promise<any> {
   docs = docs.filter(notNull);
   if (!docs.length) return;
   const wsFolder = vscode.workspace.getWorkspaceFolder(docs[0].uri);
-  const flags =
-    flagsOverride ??
-    vscode.workspace.getConfiguration("objectscript", wsFolder || docs[0].uri).get<string>("compileFlags");
+  const defaultFlags = vscode.workspace
+    .getConfiguration("objectscript", wsFolder || docs[0].uri)
+    .get<string>("compileFlags");
+  const flags = askFlags
+    ? await vscode.window.showInputBox({ title: "Enter compilation flags and qualifiers", value: defaultFlags })
+    : defaultFlags;
+  if (flags == undefined) return; // User clicked out of input box
   const api = new AtelierAPI(docs[0].uri);
   const docNames = docs.map((d) => d.name);
   // Determine the line ending to use for other documents affected
@@ -344,8 +423,8 @@ export async function compile(docs: (CurrentTextFile | CurrentBinaryFile)[], fla
             }
             return docs;
           })
-          .catch(() => {
-            compileErrorMsg();
+          .catch((error) => {
+            compileErrorMsg(error);
             // Always fetch server changes, even when compile failed or got cancelled
             return docs;
           })
@@ -370,19 +449,16 @@ export async function importAndCompile(document?: vscode.TextDocument, askFlags 
     return;
   }
 
-  const defaultFlags = config().compileFlags;
-  const flags = askFlags ? await compileFlags() : defaultFlags;
-  return importFile(file)
-    .catch((error) => {
-      throw error;
-    })
-    .then(() => {
-      if (isCompilable(file.name)) {
-        return compile([file], flags);
-      }
-
-      return undefined;
-    });
+  try {
+    if (isCompilable(file.name)) {
+      await importFile(file, true);
+      compile([file], askFlags);
+    } else {
+      await importFile(file, false);
+    }
+  } catch {
+    // importFile handles any server errors
+  }
 }
 
 export async function compileOnly(document?: vscode.TextDocument, askFlags = false): Promise<any> {
@@ -401,10 +477,7 @@ export async function compileOnly(document?: vscode.TextDocument, askFlags = fal
     return;
   }
 
-  if (isCompilable(file.name)) {
-    const flags = askFlags ? await compileFlags() : undefined;
-    compile([file], flags);
-  }
+  if (isCompilable(file.name)) compile([file], askFlags);
 }
 
 // Compiles all files types in the namespace
@@ -455,28 +528,26 @@ async function importFiles(files: vscode.Uri[], noCompile = false) {
   await Promise.allSettled<void>(
     files.map((uri) =>
       rateLimiter.call(async () => {
-        return vscode.workspace.fs
-          .readFile(uri)
-          .then((contentBytes) =>
-            currentFileFromContent(
-              uri,
-              isText(uri.path.split("/").pop(), Buffer.from(contentBytes))
-                ? new TextDecoder().decode(contentBytes)
-                : Buffer.from(contentBytes)
-            )
-          )
-          .then((curFile) => {
-            if (curFile) {
-              if (typeof curFile.content == "string" && isCompilable(curFile.name)) toCompile.push(curFile);
-              return importFile(curFile).then(() => outputChannel.appendLine("Imported file: " + curFile.fileName));
-            }
-          });
+        const contentBytes = await vscode.workspace.fs.readFile(uri);
+        const curFile = currentFileFromContent(
+          uri,
+          isText(uri.path.split("/").pop(), Buffer.from(contentBytes))
+            ? new TextDecoder().decode(contentBytes)
+            : Buffer.from(contentBytes)
+        );
+        if (curFile) {
+          if (typeof curFile.content == "string" && isCompilable(curFile.name)) {
+            toCompile.push(curFile);
+          }
+          await importFile(curFile, !noCompile);
+          outputChannel.appendLine("Imported file: " + curFile.fileName);
+        }
       })
     )
   );
 
   if (!noCompile && toCompile.length > 0) {
-    return compile(toCompile);
+    await compile(toCompile);
   }
   return;
 }
@@ -536,59 +607,8 @@ export async function compileExplorerItems(nodes: NodeBase[]): Promise<any> {
             throw new Error(`${info}Compile error`);
           }
         })
-        .catch(() => compileErrorMsg())
+        .catch((error) => compileErrorMsg(error))
   );
-}
-
-/** Import file `name` to server `api`. Used for importing local files that are not used as part of a client-side editing workspace. */
-async function importFileFromContent(
-  name: string,
-  content: string,
-  api: AtelierAPI,
-  ignoreConflict?: boolean,
-  skipDeplCheck = false
-): Promise<void> {
-  if (name.split(".").pop().toLowerCase() === "cls" && !skipDeplCheck) {
-    if (await isClassDeployed(name, api)) {
-      vscode.window.showErrorMessage(`Cannot import ${name} because it is deployed on the server.`, "Dismiss");
-      return Promise.reject();
-    }
-  }
-  ignoreConflict = ignoreConflict || config("overwriteServerChanges");
-  return api
-    .putDoc(
-      name,
-      {
-        content: content.split(/\r?\n/),
-        enc: false,
-        // We don't have an mtime for this file because it's outside a local workspace folder
-        mtime: 0,
-      },
-      ignoreConflict
-    )
-    .then(() => {
-      return;
-    })
-    .catch((error) => {
-      if (error?.statusCode == 409) {
-        return vscode.window
-          .showErrorMessage(
-            `Failed to import '${name}' because it already exists on the server. Overwrite server copy?`,
-            "Yes",
-            "No"
-          )
-          .then((action) => {
-            if (action == "Yes") {
-              return importFileFromContent(name, content, api, true, true);
-            } else {
-              return Promise.reject();
-            }
-          });
-      } else {
-        handleError(error, `Failed to save file '${name}' on the server.`);
-        return Promise.reject();
-      }
-    });
 }
 
 /** Prompt the user to compile documents after importing them */
@@ -623,7 +643,7 @@ async function promptForCompile(imported: string[], api: AtelierAPI, isIsfs: boo
                     throw new Error(`${info}Compile error`);
                   }
                 })
-                .catch(() => compileErrorMsg())
+                .catch((error) => compileErrorMsg(error))
                 .finally(() => {
                   if (isIsfs) {
                     // Refresh the files explorer to show the new files
@@ -644,207 +664,57 @@ async function promptForCompile(imported: string[], api: AtelierAPI, isIsfs: boo
   }
 }
 
-/** Import files from the local file system into a server-namespace from an `isfs` workspace folder. */
-export async function importLocalFilesToServerSideFolder(wsFolderUri: vscode.Uri): Promise<any> {
-  if (
-    !(
-      wsFolderUri instanceof vscode.Uri &&
-      wsFolderUri.scheme == FILESYSTEM_SCHEMA &&
-      (vscode.workspace.workspaceFolders != undefined
-        ? vscode.workspace.workspaceFolders.some((wsFolder) => wsFolder.uri.toString() == wsFolderUri.toString())
-        : false)
-    )
-  ) {
-    // Need an isfs workspace folder URI
-    return;
-  }
-  if (vscode.workspace.workspaceFile.scheme != "file") {
-    vscode.window.showErrorMessage(
-      "'Import Local Files...' command is not supported for unsaved workspaces.",
-      "Dismiss"
-    );
-    return;
-  }
-  const api = new AtelierAPI(wsFolderUri);
-  // Get the default URI and remove the file anme
-  let defaultUri = lastUsedLocalUri() ?? vscode.workspace.workspaceFile;
-  defaultUri = defaultUri.with({ path: defaultUri.path.split("/").slice(0, -1).join("/") });
-  // Prompt the user for files to import
-  let uris = await vscode.window.showOpenDialog({
-    canSelectFiles: true,
-    canSelectFolders: false,
-    canSelectMany: true,
-    openLabel: "Import",
-    filters: {
-      "InterSystems Files": ["cls", "mac", "int", "inc"],
-    },
-    // Need a default URI with file scheme or the open dialog
-    // will show the virtual files from the workspace folder
-    defaultUri,
-  });
-  if (!Array.isArray(uris) || uris.length == 0) {
-    // No files to import
-    return;
-  }
-  // Filter out non-ISC files
-  uris = uris.filter((uri) => isClassOrRtn(uri.path));
-  if (uris.length == 0) {
-    vscode.window.showErrorMessage("No classes or routines were selected.", "Dismiss");
-    return;
-  }
-  lastUsedLocalUri(uris[0]);
-  // Get the name and content of the files to import
-  const textDecoder = new TextDecoder();
-  const docs = await Promise.allSettled<{ name: string; content: string; uri: vscode.Uri }>(
-    uris.map((uri) =>
-      vscode.workspace.fs
-        .readFile(uri)
-        .then((contentBytes) => textDecoder.decode(contentBytes))
-        .then((content) => {
-          // Determine the name of this file
-          let docName = "";
-          let ext = "";
-          if (uri.path.split(".").pop().toLowerCase() == "cls") {
-            // Allow Unicode letters
-            const match = content.match(classNameRegex);
-            if (match) {
-              [, docName, ext = "cls"] = match;
-            }
-          } else {
-            const match = content.match(routineNameTypeRegex);
-            if (match) {
-              [, docName, ext = "mac"] = match;
-            } else {
-              const basename = uri.path.split("/").pop();
-              docName = basename.slice(0, basename.lastIndexOf("."));
-              ext = basename.slice(basename.lastIndexOf(".") + 1);
-            }
-          }
-          if (docName != "" && ext != "") {
-            return {
-              name: `${docName}.${ext.toLowerCase()}`,
-              content,
-              uri,
-            };
-          } else {
-            return Promise.reject();
-          }
-        })
-    )
-  ).then((results) => results.map((result) => (result.status == "fulfilled" ? result.value : null)).filter(notNull));
-  // The user is importing into a server-side folder, so fire the import list User Action
-  const docNames = docs.map((e) => e.name).join(",");
-  await new StudioActions().fireImportUserAction(api, docNames);
-  // Check the status of the documents to be imported and skip any that are read-only
-  await api.actionQuery("select * from %Atelier_v1_Utils.Extension_GetStatus(?)", [docNames]).then((data) =>
-    data?.result?.content?.forEach((e) => {
-      if (!e.editable) {
-        const idx = docs.findIndex((d) => {
-          const nameSplit = d.name.split(".");
-          return e.name == `${nameSplit.slice(0, -1).join(".")}.${nameSplit.pop().toUpperCase()}`;
-        });
-        if (idx != -1) {
-          docs.splice(idx, 1);
-          outputChannel.appendLine(
-            `Skipping '${e.name}' because it has been marked read-only by server-side source control.`
-          );
-        }
-      }
-    })
-  );
-  // Import the files
-  const rateLimiter = new RateLimiter(50);
-  return Promise.allSettled<string>(
-    docs.map((doc) =>
-      rateLimiter.call(async () => {
-        // Allow importing over deployed classes since the XML import
-        // command and SMP, terminal, and Studio imports allow it
-        return importFileFromContent(doc.name, doc.content, api, false, true).then(() => {
-          outputChannel.appendLine("Imported file: " + doc.uri.path.split("/").pop());
-          return doc.name;
-        });
-      })
-    )
-  ).then((results) =>
-    promptForCompile(
-      results.map((result) => (result.status == "fulfilled" ? result.value : null)).filter(notNull),
-      api,
-      true
-    )
-  );
-}
-
-interface XMLQuickPickItem extends vscode.QuickPickItem {
+interface FileQuickPickItem extends vscode.QuickPickItem {
   file: string;
 }
 
-export async function importXMLFiles(): Promise<any> {
+export async function importArbitraryFiles(): Promise<any> {
   try {
-    // Use the server connection from a workspace folder
-    const wsFolder = await getWsFolder(
-      "Pick a workspace folder. Server-side folders import from the local file system.",
-      false,
-      false,
-      false,
-      true
-    );
-    if (!wsFolder) {
-      if (wsFolder === undefined) {
+    const wsFolderUri = await getWsServerConnection("2023.2.0");
+    if (!wsFolderUri) {
+      if (wsFolderUri === undefined) {
         // Strict equality needed because undefined == null
         vscode.window.showErrorMessage(
-          "'Import XML Files...' command requires a workspace folder with an active server connection.",
+          "'Import Files...' command requires an active server connection to InterSystems IRIS version 2023.2 or above.",
           "Dismiss"
         );
       }
       return;
     }
-    const api = new AtelierAPI(wsFolder.uri);
-    // Make sure the server has the xml endpoints
-    if (api.config.apiVersion < 7) {
-      vscode.window.showErrorMessage(
-        "'Import XML Files...' command requires InterSystems IRIS version 2023.2 or above.",
-        "Dismiss"
-      );
-      return;
-    }
-    let defaultUri = wsFolder.uri;
+    const api = new AtelierAPI(wsFolderUri);
+    let defaultUri = wsFolderUri;
     if (defaultUri.scheme == FILESYSTEM_SCHEMA) {
       // Need a default URI without the isfs scheme or the open dialog
       // will show the server-side files instead of local ones
       defaultUri = lastUsedLocalUri() ?? vscode.workspace.workspaceFile;
       if (defaultUri.scheme != "file") {
-        vscode.window.showErrorMessage(
-          "'Import XML Files...' command is not supported for unsaved workspaces.",
-          "Dismiss"
-        );
+        vscode.window.showErrorMessage("'Import Files...' command is not supported for unsaved workspaces.", "Dismiss");
         return;
       }
       // Remove the file name from the URI
       defaultUri = defaultUri.with({ path: defaultUri.path.split("/").slice(0, -1).join("/") });
     }
     // Prompt the user the file to import
+    const supportedExts = ["cls", "mac", "int", "inc", "xml", "rtn", "ro"];
     let uris = await vscode.window.showOpenDialog({
       canSelectFiles: true,
       canSelectFolders: false,
       canSelectMany: true,
       openLabel: "Import",
       filters: {
-        "XML Files": ["xml"],
+        "InterSystems Files": supportedExts,
       },
       defaultUri,
     });
-    if (!Array.isArray(uris) || uris.length == 0) {
-      // No file to import
-      return;
-    }
-    // Filter out non-XML files
-    uris = uris.filter((uri) => uri.path.split(".").pop().toLowerCase() == "xml");
+    if (!uris?.length) return;
+    // Filter out non-importable files
+    uris = uris.filter((uri) => supportedExts.includes(uri.path.split(".").pop().toLowerCase()));
     if (uris.length == 0) {
-      vscode.window.showErrorMessage("No XML files were selected.", "Dismiss");
+      vscode.window.showErrorMessage("No selected files are importable.", "Dismiss");
       return;
     }
     lastUsedLocalUri(uris[0]);
-    // Read the XML files
+    // Read the files
     const fileTimestamps: Map<string, string> = new Map();
     const filesToList = await Promise.allSettled(
       uris.map(async (uri) => {
@@ -857,11 +727,23 @@ export async function importXMLFiles(): Promise<any> {
           content: new TextDecoder().decode(await vscode.workspace.fs.readFile(uri)).split(/\r?\n/),
         };
       })
-    ).then((results) => results.map((result) => (result.status == "fulfilled" ? result.value : null)).filter(notNull));
+    ).then((results) =>
+      results
+        .map((result) => {
+          if (result.status == "fulfilled") {
+            return result.value;
+          } else {
+            handleError(result.reason);
+            return null;
+          }
+        })
+        .filter(notNull)
+    );
     if (filesToList.length == 0) {
+      vscode.window.showErrorMessage("Failed to read the text of every selected file.", "Dismiss");
       return;
     }
-    // List the documents in the XML files
+    // List the documents in the files
     const documentsPerFile = await api.actionXMLList(filesToList).then((data) => data.result.content);
     // Prompt the user to select documents to import
     const quickPickItems = documentsPerFile
@@ -874,7 +756,7 @@ export async function importXMLFiles(): Promise<any> {
         }
       })
       .flatMap((file) => {
-        const items: XMLQuickPickItem[] = [];
+        const items: FileQuickPickItem[] = [];
         if (file.documents.length > 0) {
           // Add a separator for this file
           items.push({
@@ -901,10 +783,9 @@ export async function importXMLFiles(): Promise<any> {
       ignoreFocusOut: true,
       title: `Select the documents to import into namespace '${api.ns}' on server '${api.serverId}'`,
     });
-    if (docsToImport == undefined || docsToImport.length == 0) {
-      return;
-    }
-    const isIsfs = filesystemSchemas.includes(wsFolder.uri.scheme);
+    if (!docsToImport?.length) return;
+    outputChannel.show(true);
+    const isIsfs = filesystemSchemas.includes(wsFolderUri.scheme);
     if (isIsfs) {
       // The user is importing into a server-side folder
       const docNames = [...new Set(docsToImport.map((qpi) => qpi.label))].join(",");
@@ -949,6 +830,6 @@ export async function importXMLFiles(): Promise<any> {
     // Prompt the user for compilation
     promptForCompile([...new Set(imported)], api, isIsfs);
   } catch (error) {
-    handleError(error, "Error executing 'Import XML Files...' command.");
+    handleError(error, "Error executing 'Import Files...' command.");
   }
 }